LayoutTests/fast/dom/shadow/drop-event-for-input-in-shadow.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/polyfill.js"></script>
<script src="../resources/event-sender-util.js"></script>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<div id="host"></div>
<pre id="console"></pre>

<script>
function createInput(name) {
    var div = document.createElement('input');
    div.style.width = '100px';
    div.style.height = '100px';

    div.addEventListener('drop', function(e) {
        debug('PASS: drop event is fired.');
    });

    return div;
}

var shadowRoot = new WebKitShadowRoot(host);
var shadowInput = createInput('shadow');
shadowRoot.appendChild(shadowInput);

eventSender.beginDragWithFiles(["../resources/apple.gif"]);
mouseMoveToElem(shadowInput);
eventSender.mouseUp();

var successfullyParsed = true;
</script>

<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
