safari-600.7-branch/LayoutTests/fast/dom/shadow/drop-event-for-input-in-shadow-expected.txt

PASS: drop event is fired.
